OneDrive/Desktop/forex-bot/requirements.txt
# OANDA API Integration
oandapyV20==0.7.2

# TradingView Webhook Integration
flask==3.1.0
flask-cors==5.0.0

# Data Analysis and Technical Indicators
pandas==2.2.3
numpy==2.2.1
ta==0.11.0

# HTTP Requests
requests==2.32.3

# Configuration Management
python-dotenv==1.0.1
pyyaml==6.0.2

# Logging
colorlog==6.9.0

# Database (Optional - for trade history)
sqlalchemy==2.0.36

# Validation
pydantic==2.10.5

# Testing
pytest==8.3.4
pytest-mock==3.14.0

# Type Hints
typing-extensions==4.12.2
